pyproject.toml
[project]
name = "pdsim-backend"
version = "1.1.0"
authors = [
  { name="Emanuele De Pellegrin", email="edepell@ed.ac.uk" },
  { name="Andy Edmondson", email="ae2016@hw.ac.uk" },
]
description = "Using Unified Planning Library to parse PDDL and generate plans for PDSim."
dependencies= [
    'pyzmq~=25.1.1',
    'protobuf',
    'unified-planning==1.2.0',
    'up-fast-downward',
    'up-tamer',
    'up-pyperplan',
    'up-enhsp',
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = { file = "./LICENSE" }

[project.urls]
Homepage = "https://github.com/Cryoscopic-E/PDSim-Backend"
Issues = "https://github.com/Cryoscopic-E/PDSim-Backend/issues"
